src/main/webapp/app/entities/pf-loan-repayment/list/pf-loan-repayment.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { PfLoanRepaymentService } from '../service/pf-loan-repayment.service';

import { PfLoanRepaymentComponent } from './pf-loan-repayment.component';
import SpyInstance = jest.SpyInstance;

describe('PfLoanRepayment Management Component', () => {
  let comp: PfLoanRepaymentComponent;
  let fixture: ComponentFixture<PfLoanRepaymentComponent>;
  let service: PfLoanRepaymentService;
  let routerNavigateSpy: SpyInstance<Promise<boolean>>;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [
        RouterTestingModule.withRoutes([{ path: 'pf-loan-repayment', component: PfLoanRepaymentComponent }]),
        HttpClientTestingModule,
      ],
      declarations: [PfLoanRepaymentComponent],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            data: of({
              defaultSort: 'id,asc',
            }),
            queryParamMap: of(
              jest.requireActual('@angular/router').convertToParamMap({
                page: '1',
                size: '1',
                sort: 'id,desc',
              })
            ),
            snapshot: { queryParams: {} },
          },
        },
      ],
    })
      .overrideTemplate(PfLoanRepaymentComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(PfLoanRepaymentComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PfLoanRepaymentService);
    routerNavigateSpy = jest.spyOn(comp.router, 'navigate');

    const headers = new HttpHeaders();
    jest.spyOn(service, 'query').mockReturnValue(
      of(
        new HttpResponse({
          body: [{ id: 123 }],
          headers,
        })
      )
    );
  });

  it('Should call load all on init', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.pfLoanRepayments?.[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  describe('trackId', () => {
    it('Should forward to pfLoanRepaymentService', () => {
      const entity = { id: 123 };
      jest.spyOn(service, 'getPfLoanRepaymentIdentifier');
      const id = comp.trackId(0, entity);
      expect(service.getPfLoanRepaymentIdentifier).toHaveBeenCalledWith(entity);
      expect(id).toBe(entity.id);
    });
  });

  it('should load a page', () => {
    // WHEN
    comp.navigateToPage(1);

    // THEN
    expect(routerNavigateSpy).toHaveBeenCalled();
  });

  it('should calculate the sort attribute for an id', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenLastCalledWith(expect.objectContaining({ sort: ['id,desc'] }));
  });

  it('should calculate the sort attribute for a non-id attribute', () => {
    // GIVEN
    comp.predicate = 'name';

    // WHEN
    comp.navigateToWithComponentValues();

    // THEN
    expect(routerNavigateSpy).toHaveBeenLastCalledWith(
      expect.anything(),
      expect.objectContaining({
        queryParams: expect.objectContaining({
          sort: ['name,asc'],
        }),
      })
    );
  });
});
